# api.yaml
openapi: 3.0.3
info:
  title: Bingo  Maker API
  description: A service for creating and managing tile pools containing text or image tiles and generating bingo cards from those pools.
  version: 0.2.0
...
paths:
...
  /bingocard/{tilepoolId}:
    get:
      summary: Generate a bingo card from a tile pool
      description: Generate a bingo card based on the tiles from an existing tile pool.
      parameters:
        - name: tilepoolId
          in: path
          required: true
          description: The ID of the tile pool to generate the bingo card from
          schema:
            type: string
        - name: size
          in: query
          required: false
          description: The size (width and height) of the bingo card (e.g., 5 for a 5x5 grid)
          schema:
            type: integer
            default: 5
        - name: excluded_tags
          in: query
          required: false
          description: A array of tags to exlude for this card
          schema:
            type: array
            default: []
            items:
              type: string
        - name: seed
          in: query
          required: false
          description: The seed to use when generating a bingo card
          schema:
            type: integer
      responses:
        '200':
          description: A JSON object representing the generated bingo card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BingoCard'
        '404':
          description: Tile pool not found
        '400':
          description: Invalid input or request parameters
      tags:
        - Bingo Cards
components:
  schemas:
...
    Tile:
      type: object
      properties:
        type:
          type: string
          enum: [text, image]
          description: The type of tile; it can either be 'text' or 'image'
        content:
          type: string
          description: The content of the tile (either a text string or image URL)
        tags:
          type: array
          description: A array containing the tags for this tile
          items:
            type: string
            description: A descriptor for the tile 
          
      required:
        - type
        - content
    BingoCard:
      type: object
      properties:
        id:
          type: string
          description: The unique ID of the generated bingo card
        tiles:
          type: array
          description: A array representing the bingo card layout with tile content.
          items:
            $ref: '#/components/schemas/Tile'
        size:
          type: integer
          description: The dimensions (size x size) of the bingo card grid
      required:
        - id
        - grid
        - size